refactor: sqlite service
